requirements.txt
streamlit>=1.37.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
feedparser>=6.0.11
python-dotenv>=1.0.1
pydantic>=2.7.0
openai>=1.50.0
